+++ src-tauri/src/responses/buvid3_resp_data.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
+pub struct Buvid3RespData {
+    pub buvid: String,
+}
