+++ src-tauri/src/responses/generate_qrcode_resp_data.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
+pub struct GenerateQrcodeRespData {
+    pub url: String,
+    #[serde(rename = "auth_code")]
+    pub auth_code: String,
+}
