+++ src-tauri/src/responses/manga_resp_data.rs
@@ -1,176 +1,5 @@
-// TODO: 把responses.rs拆成多个文件
 use serde::{Deserialize, Serialize};
 use specta::Type;
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BiliResp {
-    pub code: i64,
-    #[serde(default, alias = "message")]
-    pub msg: String,
-    pub data: Option<serde_json::Value>,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct GenerateQrcodeRespData {
-    pub url: String,
-    #[serde(rename = "auth_code")]
-    pub auth_code: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct QrcodeStatusRespData {
-    #[serde(rename = "is_new")]
-    pub is_new: bool,
-    pub mid: i64,
-    #[serde(rename = "access_token")]
-    pub access_token: String,
-    #[serde(rename = "refresh_token")]
-    pub refresh_token: String,
-    #[serde(rename = "expires_in")]
-    pub expires_in: i64,
-    #[serde(rename = "token_info")]
-    pub token_info: TokenInfoRespData,
-    #[serde(rename = "cookie_info")]
-    pub cookie_info: CookieInfoRespData,
-    pub sso: Vec<String>,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct TokenInfoRespData {
-    pub mid: i64,
-    #[serde(rename = "access_token")]
-    pub access_token: String,
-    #[serde(rename = "refresh_token")]
-    pub refresh_token: String,
-    #[serde(rename = "expires_in")]
-    pub expires_in: i64,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct CookieInfoRespData {
-    pub cookies: Vec<CookieRespData>,
-    pub domains: Vec<String>,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct CookieRespData {
-    pub name: String,
-    pub value: String,
-    #[serde(rename = "http_only")]
-    pub http_only: i64,
-    pub expires: i64,
-    pub secure: i64,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct Buvid3RespData {
-    pub buvid: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct SearchMangaRespData {
-    pub list: Vec<MangaInSearchRespData>,
-    #[serde(rename = "total_page")]
-    pub total_page: i32,
-    #[serde(rename = "total_num")]
-    pub total_num: i32,
-    pub recommends: Vec<RecommendRespData>,
-    pub similar: String,
-    #[serde(rename = "se_id")]
-    pub se_id: String,
-    pub banner: BannerRespData,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct MangaInSearchRespData {
-    pub id: i32,
-    pub title: String,
-    #[serde(rename = "org_title")]
-    pub org_title: String,
-    #[serde(rename = "horizontal_cover")]
-    pub horizontal_cover: String,
-    #[serde(rename = "square_cover")]
-    pub square_cover: String,
-    #[serde(rename = "vertical_cover")]
-    pub vertical_cover: String,
-    #[serde(rename = "author_name")]
-    pub author_name: Vec<String>,
-    pub styles: Vec<String>,
-    #[serde(rename = "is_finish")]
-    pub is_finish: i32,
-    #[serde(rename = "allow_wait_free")]
-    pub allow_wait_free: bool,
-    #[serde(rename = "discount_type")]
-    pub discount_type: i32,
-    #[serde(rename = "type")]
-    pub type_field: i32,
-    pub wiki: WikiRespData,
-    pub numbers: i32,
-    #[serde(rename = "jump_value")]
-    pub jump_value: String,
-    #[serde(rename = "real_title")]
-    pub real_title: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct WikiRespData {
-    pub id: i32,
-    pub title: String,
-    #[serde(rename = "origin_title")]
-    pub origin_title: String,
-    #[serde(rename = "vertical_cover")]
-    pub vertical_cover: String,
-    pub producer: String,
-    #[serde(rename = "author_name")]
-    pub author_name: Vec<String>,
-    #[serde(rename = "publish_time")]
-    pub publish_time: String,
-    pub frequency: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct RecommendRespData {
-    pub id: i32,
-    pub title: String,
-    #[serde(rename = "horizontal_cover")]
-    pub horizontal_cover: String,
-    #[serde(rename = "square_cover")]
-    pub square_cover: String,
-    #[serde(rename = "vertical_cover")]
-    pub vertical_cover: String,
-    #[serde(rename = "last_short_title")]
-    pub last_short_title: String,
-    pub recommendation: String,
-    #[serde(rename = "is_finish")]
-    pub is_finish: i32,
-    pub total: i32,
-    #[serde(rename = "allow_wait_free")]
-    pub allow_wait_free: bool,
-    #[serde(rename = "author_name")]
-    pub author_name: Vec<String>,
-    pub styles: Vec<String>,
-    #[serde(rename = "discount_type")]
-    pub discount_type: i32,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct BannerRespData {
-    pub icon: String,
-    pub title: String,
-    pub url: String,
-}
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
@@ -463,74 +292,4 @@ pub struct IncreaseRespData {
     pub days: i64,
     #[serde(rename = "increase_percent")]
     pub increase_percent: i64,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ImageIndexRespData {
-    pub host: String,
-    pub images: Vec<ImageRespData>,
-    #[serde(rename = "last_modified")]
-    pub last_modified: String,
-    pub path: String,
-    pub video: VideoRespData,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ImageRespData {
-    pub path: String,
-    #[serde(rename = "video_path")]
-    pub video_path: String,
-    #[serde(rename = "video_size")]
-    pub video_size: String,
-    pub x: i64,
-    pub y: i64,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct VideoRespData {
-    #[serde(rename = "bin_url")]
-    pub bin_url: String,
-    pub filename: String,
-    #[serde(rename = "img_urls")]
-    pub img_urls: Vec<serde_json::Value>,
-    #[serde(rename = "img_x_len")]
-    pub img_x_len: i64,
-    #[serde(rename = "img_x_size")]
-    pub img_x_size: i64,
-    #[serde(rename = "img_y_len")]
-    pub img_y_len: i64,
-    #[serde(rename = "img_y_size")]
-    pub img_y_size: i64,
-    #[serde(rename = "raw_height")]
-    pub raw_height: String,
-    #[serde(rename = "raw_rotate")]
-    pub raw_rotate: String,
-    #[serde(rename = "raw_width")]
-    pub raw_width: String,
-    pub resource: Vec<serde_json::Value>,
-    pub route: String,
-    pub svid: String,
-}
-
-pub type ImageTokenRespData = Vec<ImageTokenItemRespData>;
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ImageTokenItemRespData {
-    #[serde(rename = "complete_url")]
-    pub complete_url: String,
-    #[serde(rename = "hit_encrpyt")]
-    pub hit_encrpyt: bool,
-    pub url: String,
-    pub token: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
-#[serde(rename_all = "camelCase")]
-pub struct UserProfileRespData {
-    pub face: String,
-    pub name: String,
 }
