+++ src-tauri/src/responses/user_profile_resp_data.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
+pub struct UserProfileRespData {
+    pub face: String,
+    pub name: String,
+}
